requirements.txt
streamlit
tensorflow-cpu
pandas
numpy
plotly
pillow